RavenDB-16865 Fix debounced name validation
